some files

+++ Type_Casting/App1/src/U.java
@@ -3,7 +3,7 @@ class U
 	public static void main(String[] args)
 	{
 		int i = 10;
-		short s = (byte) i;
+		short s = (short) i;
 		System.out.println("done");
 	}
 }
